% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GLS_ADF.R
\name{GLS_ADF}
\alias{GLS_ADF}
\title{Elliott, Rothenberg ve Stock(1996) GLS detrending in ADF unit root test function}
\usage{
GLS_ADF(data_name, case, max_lag, lsm)
}
\arguments{
\item{data_name}{series name,}

\item{case}{if demeaned data 1 if detrended data 2,}

\item{max_lag}{maximum lag}

\item{lsm}{lag selection methods if 1 AIC, if 2 BIC}
}
\value{
"Model" Estimated model

"Selected lag" the lag order

"Test Statistic" the value of the test statistic
}
\description{
This function allows you to make Elliott, Rothenberg ve Stock(1996) GLS detrending method for the unit root test procedure developed by Dickey and Fuller(1981).
}
\examples{

x <- rnorm(1000)
GLS_ADF(x, case = 1, lags = 6, lsm =1)


y <- cumsum(rnorm(1000))
GLS_ADF(y, 1, 3, 3)



}
\references{
Elliott, G., T. J. Rothenberg, and J. H. Stock. 1996. Efficient tests for an autoregressive unit root. Econometrica 64 (4):813–36.


Burak Guris, R Uygulamalı Dogrusal Olmayan Zaman Serileri Analizi, DER Yayinevi, 2020.
}
\keyword{GLS}
\keyword{nonlinear}
\keyword{root}
\keyword{test}
\keyword{unit}
