% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Guris_Guris_2022.R
\name{Guris_Guris_2022}
\alias{Guris_Guris_2022}
\title{Guris and Guris(2022) GLS detrending in nonlinear unit root test function}
\usage{
Guris_Guris_2022(data_name, case, lags, lsm)
}
\arguments{
\item{data_name}{series name,}

\item{case}{if demeaned data 1 if detrended data 2,}

\item{lags}{maximum lag}

\item{lsm}{lag selection methods if 1 AIC, if 2 BIC, if 3 t-stat significance}
}
\value{
"Model" Estimated model

"Selected lag" the lag order

"Test Statistic" the value of the test statistic
}
\description{
This function allows you to make Güriş and Güriş GLS detrending method for the unit root test procedure developed by Kruse 2011. A new unit root
test against ESTAR based on a class of modified statistics. Statistical Papers 52 (1):71–85.
}
\examples{

x <- rnorm(1000)
Guris_Guris_2022(x, case = 1, lags = 6, lsm =1)


y <- cumsum(rnorm(1000))
Guris_Guris_2022(y, 1, 3, 3)



}
\references{
Kruse 2011. A new unit root test against ESTAR based on a class of modified statistics. Statistical Papers 52 (1):71–85.


Burak Guris, R Uygulamalı Dogrusal Olmayan Zaman Serileri Analizi, DER Yayinevi, 2020.
}
\keyword{GLS}
\keyword{nonlinear}
\keyword{root}
\keyword{test}
\keyword{unit}
