% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Kapetanios_Shin_2008.R
\name{Kapetanios_Shin_2008}
\alias{Kapetanios_Shin_2008}
\title{Kapetanios and Shin(2008) GLS detrending in KSS(2003) nonlinear unit root test function}
\usage{
Kapetanios_Shin_2008(data_name, case, lags, lsm)
}
\arguments{
\item{data_name}{series name,}

\item{case}{if demeaned data 1 if detrended data 2,}

\item{lags}{maximum lag}

\item{lsm}{lag selection methods if 1 AIC, if 2 BIC, if 3 t-stat significance}
}
\value{
"Model" Estimated model

"Selected lag" the lag order

"Test Statistic" the value of the test statistic
}
\description{
This function allows you to make Kapetanios and Shin(2008) GLS detrending method for the unit root test procedure developed by Kapetanios, Shin and Snell(2003).
}
\examples{

x <- rnorm(1000)
Kapetanios_Shin_2008(x, case = 1, lags = 6, lsm =1)


y <- cumsum(rnorm(1000))
Kapetanios_Shin_2008(y, 2, 12, 2)



}
\references{
Kapetanios, G., Shin, Y., & Snell, A. (2003). Testing for a unit root in the nonlinear STAR framework. Journal of econometrics, 112(2), 359-379.


Burak Guris, R Uygulamalı Dogrusal Olmayan Zaman Serileri Analizi, DER Yayinevi, 2020.
}
\keyword{GLS}
\keyword{nonlinear}
\keyword{root}
\keyword{test}
\keyword{unit}
