% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Su_Nguen_2013.R
\name{Su_Nguyen_2013}
\alias{Su_Nguyen_2013}
\title{Su and Nguyen(2013) GLS detrending in Sollis nonlinear unit root test function}
\usage{
Su_Nguyen_2013(data_name, case, lags, lsm)
}
\arguments{
\item{data_name}{series name,}

\item{case}{if demeaned data 1 if detrended data 2,}

\item{lags}{maximum lag}

\item{lsm}{lag selection methods if 1 AIC, if 2 BIC, if 3 t-stat significance}
}
\value{
"Model" Estimated model

"Selected lag" the lag order

"Test Statistic" the value of the test statistic
}
\description{
This function allows you to make Su and Nguyen(2013) GLS detrending method for the unit root test procedure developed by Kruse 2009.
}
\examples{

x <- rnorm(1000)
Su_Nguyen_2013(x, case = 2, lags = 3, lsm = 2)


y <- cumsum(rnorm(1000))
Su_Nguyen_2013(y, 1, 12, 1)



}
\references{
Jen-Je Su & Jeremy K. Nguyen (2013) GLS detrending in Sollis nonlinear unit root tests, Applied Economics Letters, 20:13, 1259-1262


Burak Guris, R Uygulamalı Dogrusal Olmayan Zaman Serileri Analizi, DER Yayinevi, 2020.
}
\keyword{GLS}
\keyword{nonlinear}
\keyword{root}
\keyword{test}
\keyword{unit}
