% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tr_inf_rate.R
\docType{data}
\name{Tr_inf_rate}
\alias{Tr_inf_rate}
\title{Tr_inf_rate}
\format{
A data frame containing :

Inflation Rate in Turkey
}
\source{
The Central Bank of the Republic of Turkey
}
\usage{
Tr_inf_rate
}
\description{
Monthly time series data between 01.2005 - 10.2023
}
\examples{
summary(Tr_inf_rate)
}
\keyword{datasets}
